+++ android/features/shopping-list/purchase-input/src/main/kotlin/io/chefbook/features/shoppinglist/purchases/input/ui/mvi/PurchaseInputDialogIntent.kt
@@ -5,7 +5,7 @@ import io.chefbook.libs.mvi.MviIntent
 
 sealed class PurchaseInputDialogIntent : MviIntent {
   data class SetName(val name: String) : PurchaseInputDialogIntent()
-  data class SetAmount(val amount: Int?) : PurchaseInputDialogIntent()
+  data class SetAmount(val amount: Float?) : PurchaseInputDialogIntent()
   data class SetMeasureUnit(val unit: MeasureUnit?) : PurchaseInputDialogIntent()
 
   data object Close : PurchaseInputDialogIntent()
