+++ common/sdk/shopping-list/api/external/src/commonMain/kotlin/io/chefbook/sdk/shoppinglist/api/external/domain/entities/Purchase.kt
@@ -5,7 +5,7 @@ import io.chefbook.libs.models.measureunit.MeasureUnit
 data class Purchase(
   val id: String,
   val name: String,
-  val amount: Int? = null,
+  val amount: Float? = null,
   val measureUnit: MeasureUnit? = null,
   val multiplier: Int = 1,
   val isPurchased: Boolean = false,
